made it so you cannot delete the only task in no category

+++ src/app/components/tasks/tasks.component.ts
@@ -135,9 +135,9 @@ refreshTaskList(){
   //
   //
   deleteTask(categoryId,task){
+     if(task.title === "Deleted task")
+       return;
     this.firebaseService.deleteTask(categoryId,task);
-
-
 
   }
